# makeJSCallback optimization and bugfixes