--- a/RootInteractive/InteractiveDrawing/bokeh/ColumnFilter.py
+++ b/RootInteractive/InteractiveDrawing/bokeh/ColumnFilter.py
@@ -1,8 +1,8 @@
-from bokeh.model import Model
+from RootInteractive.InteractiveDrawing.bokeh.RIFilter import RIFilter
 from bokeh.core.properties import Instance, String
 from bokeh.models.sources import ColumnarDataSource
 
-class ColumnFilter(Model):
+class ColumnFilter(RIFilter):
     __implementation__ = "ColumnFilter.ts"
 
     source = Instance(ColumnarDataSource, help="Column data source to select from")