--- a/RootInteractive/InteractiveDrawing/bokeh/ColumnFilter.ts
+++ b/RootInteractive/InteractiveDrawing/bokeh/ColumnFilter.ts
@@ -53,7 +53,8 @@ export class ColumnFilter extends RIFilter {
     if (!dirty_source){
         return cached_vector
     }
-    let col = source.get_array(field)
+    let col = source.get_column(field)
+    if(col == null) return []
     let new_vector: boolean[] = this.cached_vector
     if (new_vector == null){
         new_vector = Array(col.length)