--- a/RootInteractive/InteractiveDrawing/bokeh/RangeFilter.py
+++ b/RootInteractive/InteractiveDrawing/bokeh/RangeFilter.py
@@ -1,8 +1,8 @@
-from bokeh.model import Model
+from RootInteractive.InteractiveDrawing.bokeh.RIFilter import RIFilter
 from bokeh.core.properties import Instance, String, Tuple, Float
 from bokeh.models.sources import ColumnarDataSource
 
-class RangeFilter(Model):
+class RangeFilter(RIFilter):
     __implementation__ = "RangeFilter.ts"
 
     source = Instance(ColumnarDataSource, help="Column data source to select from")