--- a/RootInteractive/InteractiveDrawing/bokeh/RangeFilter.ts
+++ b/RootInteractive/InteractiveDrawing/bokeh/RangeFilter.ts
@@ -1,11 +1,11 @@
 import {ColumnarDataSource} from "models/sources/columnar_data_source"
-import {Model} from "model"
+import {RIFilter} from "./RIFilter"
 import * as p from "core/properties"
 
 export namespace RangeFilter {
   export type Attrs = p.AttrsOf<Props>
 
-  export type Props = Model.Props & {
+  export type Props = RIFilter.Props & {
     source: p.Property<ColumnarDataSource>
     field: p.Property<string>
     range: p.Property<number[]>
@@ -14,7 +14,7 @@ export namespace RangeFilter {
 
 export interface RangeFilter extends RangeFilter.Attrs {}
 
-export class RangeFilter extends Model {
+export class RangeFilter extends RIFilter {
   properties: RangeFilter.Props
 
   constructor(attrs?: Partial<RangeFilter.Attrs>) {