--- a/RootInteractive/InteractiveDrawing/bokeh/RangeFilter.ts
+++ b/RootInteractive/InteractiveDrawing/bokeh/RangeFilter.ts
@@ -68,7 +68,7 @@ export class RangeFilter extends RIFilter {
     if (!dirty_source && !dirty_widget){
         return cached_vector
     }
-    let col = source.get_array(field) as number[]
+    let col = source.get_column(field) as number[]
     let new_vector: boolean[] = this.cached_vector
     if (new_vector == null){
         new_vector = Array(col.length)