This PR:
- Fixes various bugs in makeJSCallback negatively affecting performance
- - test_bokehClientHistogram.py - 5 widgets, 5e6 points, time for selecting points went from 300ms to 60ms
- - ATO-609/trackClusterDumpDraw.ipynb -  3e6 points, 8 widgets, time for selecting points went from 240ms to 60ms
- Fixes a bug with plot title when using custom axis titles
- Fixes a crash that sometimes happens when downsampler is initialized (couldn't reliably reproduce)